requirements.txt
Flask==2.3.2
boto3==1.34.0
python-dotenv==1.0.0
awscli==1.32.0
requests==2.31.0
pytest==7.4.0